refactor: base 패키지 의존성 정리
