+++ src/main/java/com/bookbla/americano/domain/notification/repository/PushAlarmLogRepository.java
@@ -1,6 +1,6 @@
-package com.bookbla.americano.base.log.repository;
+package com.bookbla.americano.domain.notification.repository;
 
-import com.bookbla.americano.base.log.repository.entity.PushAlarmLog;
+import com.bookbla.americano.domain.notification.repository.entity.PushAlarmLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PushAlarmLogRepository extends JpaRepository<PushAlarmLog, Long> {
